[EasyCodingStandard] use Symfony DI [wip]
