+++ packages/EasyCodingStandard/composer.json
@@ -10,9 +10,7 @@
         "symfony/console": "^3.3",
         "symfony/dependency-injection": "^3.3",
         "symfony/finder": "^3.3",
-        "nette/bootstrap": "^2.4|^3.0",
         "nette/caching": "^2.4|^3.0",
-        "nette/di": "^2.4|^3.0",
         "nette/robot-loader": "^3.0",
         "nette/utils": "^2.4|^3.0",
         "nette/neon": "^2.4|^3.0",
