--- packages/EasyCodingStandard/packages/SniffRunner/src/DI/SniffRunnerExtension.php
@@ -1,88 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Symplify\EasyCodingStandard\SniffRunner\DI;
-
-use Nette\DI\CompilerExtension;
-use Nette\DI\ServiceDefinition;
-use Nette\Utils\Strings;
-use Symplify\EasyCodingStandard\Configuration\Option\SniffsOption;
-
-final class SniffRunnerExtension extends CompilerExtension
-{
-    public function loadConfiguration(): void
-    {
-        $sniffs = $this->getContainerBuilder()->parameters[SniffsOption::NAME];
-        $this->registerSniffsAsServices($sniffs);
-    }
-
-    /**
-     * @param mixed[] $sniffs
-     */
-    private function registerSniffsAsServices(array $sniffs): void
-    {
-        $containerBuilder = $this->getContainerBuilder();
-        foreach ($sniffs as $sniffClass => $configuration) {
-            $sniffDefinition = $this->createSniffDefinition($sniffClass, $configuration);
-            $containerBuilder->addDefinition(md5($sniffClass), $sniffDefinition);
-        }
-    }
-
-    /**
-     * @param mixed[] $configuration
-     */
-    private function createSniffDefinition(string $sniffClass, array $configuration): ServiceDefinition
-    {
-        $sniffDefinition = new ServiceDefinition;
-        $sniffDefinition->setClass($sniffClass);
-
-        foreach ($configuration as $property => $value) {
-            $sniffDefinition->addSetup(
-                '$' . $property,
-                [$this->escapeValue($value)]
-            );
-        }
-
-        return $sniffDefinition;
-    }
-
-    /**
-     * @param mixed $value
-     * @return mixed
-     */
-    private function escapeValue($value)
-    {
-        if (is_array($value)) {
-            return $this->escapeAtSignInArray($value);
-        }
-
-        if (is_string($value)) {
-            return $this->escapeAtSign($value);
-        }
-
-        return $value;
-    }
-
-    /**
-     * @param mixed[] $value
-     * @return mixed[]
-     */
-    private function escapeAtSignInArray(array $value): array
-    {
-        foreach ($value as $key => $subValue) {
-            if (is_string($subValue)) {
-                $value[$key] = $this->escapeAtSign($subValue);
-            }
-        }
-
-        return $value;
-    }
-
-    private function escapeAtSign(string $value): string
-    {
-        if (Strings::startsWith($value, '@')) {
-            return '@' . $value;
-        }
-
-        return $value;
-    }
-}
