+++ packages/EasyCodingStandard/src/config/services.yml
@@ -5,7 +5,7 @@ services:
     # PSR-4 autodiscovery
     Symplify\EasyCodingStandard\:
         resource: '../../src'
-        exclude: '../../src/{Contract,Exception,Error/Error.php,Application/Command,DependencyInjection}'
+        exclude: '../../src/{Contract,Exception,Error/Error.php,Application/Command,DependencyInjection,Configuration/Loader/NeonLoader.php}'
     Symplify\EasyCodingStandard\ChangedFilesDetector\:
         resource: '../../packages/ChangedFilesDetector/src'
         exclude: '../../packages/ChangedFilesDetector/src/{Cache/CacheFactory}'
@@ -19,6 +19,7 @@ services:
         resource: '../../packages/FixerRunner/src'
         exclude: '../../packages/FixerRunner/src/{Exception}'
 
+    Symfony\Component\Console\Application: ~
     Symfony\Component\Console\Terminal: ~
     Symfony\Component\Console\Input\ArgvInput: ~
     Symfony\Component\Console\Output\ConsoleOutput: ~
